requirements.txt
python-dotenv
openai
pandas
streamlit
